requirements.txt
# Vision X - Unified Road Analysis System Requirements
# Python 3.8+ required
# Install with: pip install -r requirements.txt

# Core Framework
streamlit>=1.28.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0

# Scientific Computing
numpy>=1.23.5
scipy>=1.10.0

# YOLOv5 Dependencies
ultralytics>=8.0.0
PyYAML>=6.0
tqdm>=4.66.3
pandas>=2.0.0
seaborn>=0.12.0

# System Utilities
psutil>=5.9.0
packaging>=23.0
setuptools>=70.0.0
urllib3>=2.0.0

# Optional GPU Support (install separately from pytorch.org)
# For CUDA 11.8: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
# For CUDA 12.1: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121

